#23 rename utils to verboseDateTime
