+++ src/Events4FriendsBotApp.js
@@ -3,7 +3,7 @@ const fetch = require("node-fetch");
 const moment = require('moment');
 require('moment/locale/ru');
 
-const verboseDateTime = require('./utils.js');
+const verboseDateTime = require('./verboseDateTime.js');
 
 const FIREBASE_DATE_FORMAT = 'YYYY-MM-DDThh:mm:ss';
 const FIREBASE_DATE_FORMAT_WITH_UTC = 'YYYY-MM-DDThh:mm:ssZZZZ';
